+++ Collaborative_Filtering/newsmf_model.py
@@ -1,5 +1,4 @@
 # Mini simple version for predictions
-
 import torch
 import torch.nn as nn
