--- Collaborative_Filtering/tb_logs/my_model/version_10/hparams.yaml
@@ -1,1 +1,0 @@
-{}
