--- Collaborative_Filtering/tb_logs/my_model/version_7/hparams.yaml
@@ -1,1 +1,0 @@
-{}
