--- Model/model_prediction.py
@@ -1,3 +1,0 @@
-# Description: This file contains the model for the application.
-def test_function(user_id):
-    return user_id;
